Improve README splitting
Small arrangements.

@a5e : I left two edge cases that aren't really resolved. I'll probably
find a way to document it somewhere else but can't find out where at
this moment.

I also moved the counters' section on the README, since it's displayed
(and useful) for everyone. I consider it as a feature of a11y.css rather
than a technical information.

And in the next days I'll add a test section on CONTRIBUTE.md.

Thanks a lot for your PR, it helped a lot!